requirements.txt
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.13.0
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.20
django-cors-headers==4.6.0
django-environ==0.12.0
django-jsonfield==1.4.1
djangorestframework==3.15.2
environ==1.0
filelock==3.17.0
fsspec==2025.2.0
h11==0.14.0
idna==3.10
imgsim==0.1.2
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.11.0.86
outcome==1.3.0.post0
pillow==11.1.0
PySocks==1.7.1
requests==2.32.3
selenium==4.28.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
sympy==1.13.3
torch==2.2.2
torchvision==0.17.2
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
opencv-python==4.11.0.86